build(debug): Replace MSVC __int64 in SimpleProfiler with long long

+++ Core/GameEngine/Source/Common/System/Debug.cpp
@@ -647,7 +647,7 @@ void SimpleProfiler::stop()
 {
 	if (m_startThisSession != 0)
 	{
-		__int64 stop;
+		long long stop;
 		QueryPerformanceCounter((LARGE_INTEGER*)&stop);
 		m_totalThisSession = stop - m_startThisSession;
 		m_totalAllSessions += stop - m_startThisSession;
